Combat Distance fixes
Fixed a bug in charge distance causing it to return the wrong value, numerous cascading problems were resolved. Several related systems were fixed in the process (grim trigger distances for lings and for scourge).  Combat is improved in test scenarioes, still does not lose to stone.

Caused the bot to check for long distance mining and building expos if it is LD mining.