Internalized unit updates.
All unit commands are followed by an update in the unit_inventory passed to the command.

The commands that are not passed a Unit_Inventory update the overarching friendly_inventory.

Sometimes there are commands which are of the first type, and contain a subroutine of the second type. In which case, the friendly_inventory and the passed unit_inventory will be updated. Perhaps this is overuse but it is superior than having an out-of-date inventory.