Player Log Update Pull Request (#128)
* Detection Attempt

The added piece of code is an attempt at detecting the time in seconds of the first enemy detector spotted.

* Detection Addition

This should be able to notify us if we have seen detection units. (Still to be tested on other peoples' bots.

* Detection Update

Refactoring and Code Improvement

* Detection Fix

* Player Data Collection

This update collects the enemy player's data from all the info provided from the Player_Model. Testing has provided results, but they are a little bit sketchy.

* PlayerLogUpdate

Added refence functionality to playerLog function.

* Player Inventory with Description

This is simply the same function but with labelled parameters.

* Player Log fix and neat formatting

Fixed an error in the player log and formatted the output file.

* test

test

* compare

compare

* f

f

* PlayerLog Update

Added new features to the playerLog function, as well as minor other changes.

* Minor Stock Fix

Fixed error where Sum of max stocks were not adding up.