Pull request created for updated write and read functions (#138)
* Detection Attempt

The added piece of code is an attempt at detecting the time in seconds of the first enemy detector spotted.

* Detection Addition

This should be able to notify us if we have seen detection units. (Still to be tested on other peoples' bots.

* Detection Update

Refactoring and Code Improvement

* Detection Fix

* Player Data Collection

This update collects the enemy player's data from all the info provided from the Player_Model. Testing has provided results, but they are a little bit sketchy.

* PlayerLogUpdate

Added refence functionality to playerLog function.

* Player Inventory with Description

This is simply the same function but with labelled parameters.

* Player Log fix and neat formatting

Fixed an error in the player log and formatted the output file.

* test

test

* compare

compare

* f

f

* PlayerLog Update

Added new features to the playerLog function, as well as minor other changes.

* ScoutingManager (#125)

* ScoutingManager v0.1

Bare-bones scouting manager, not ready for production

- Sets, clears, and updates zergling/overlord scouts

-Shouldn't overlord scout against terran
-Currently only scouts initial base locations
-Has some problems updating scouts

* Should have been included in last commit

Forgot to change an inventory to map_inventory

* ScoutingManager v0.11

Initial overlord scout is functional again (except against Terran)
Some stylistic changes
Few more things that should have been included in v0.1

* ScoutingManager v.9

-- Scouting manager now has a suicide ling scout and an expo ling scout which checks random expansions on map all game
-- Overlords should stop scouting once the enemy has units that can kill it (not fully tested)
-- Map_Inventory::MeanEnemyBuildingLocation (ignores Resource depots and Refineries)
-- Added functionality to Scouting getScoutTargets
-- Cleaned up Scouting needScout
-- Fixed Scouting updateScouts (checks for nullpointers prior now)

* ScoutingManager v9.1

-- Scouting now gets disabled when we are under threat of losing or if our army greatly outnumbers the enemy.
-- Enemy base scout now runs to the nearest mean enemy building instead of just the mean position

* ScoutingManager v.92

-- Added scout phase_ in Draw Mode for scouting units
-- Added Scouting boolean to drawTextScreen in Draw Mode
-- Readded the Game Seed to drawTextScreen in Draw Mode

* ScoutingManager v1

-- Expo ling scouts now only scout the bases closer to the enemy location. Not a fantastic implementation but was having difficulties implementing with weighted probabilities

* ScoutingManager v1.0

Potential Exception Error? Just in case

* ScoutingManager v1.01

-- Syntactic changes
-- Made a function in the Player Model to handle all the starting game conditions
-- Moved the disablement of overlord scouting against Terran to the gameStartConditions function in the Player Model

* ScoutingManager v1.01a

-- Should have been included in v1.01

* Revert gameStartConditions playerModelManager

The changes weren't working as intended and causing crashes. Reverting for now, will try to get it to work and recommit later

* ScoutingManager v1.02

-- Put in the fix for the "fall through" if-statements in the getScoutTargets function that was causing freezing
-- Untabified all my edits/changes to keep project consistency

* ScoutingManager v1.03

Added enemy_race_ variable to the player model. Based letting overlord scouts on that variable.

* ScoutingManager v1.03a

Forgot to remove diagnostic text

* ScoutingManager v1.1

-- Replaced the getScoutTargets expo ling vectors with a map of <int(base_distance), Position>
-- Complete overhaul of the getScoutTargets expo-ling logic, ScoutingManager.cpp: lines 65-114. Working okay, but not flawlessly (using the mean of radial distances is kind of wonky)
-- To avoid any confusion, a scout unit's phase_ now get's set to "None" when the scout is cleared.
-- Added a diagnostic line for scouts in DRAW_MODE (suicide zergling and overlords are purple, expo lings are yellow)
-- Moved the let_overlords_scout bool check out of the "for every unit" loop.
-- Scouting is now always enabled by default.
-- Added the clock timer to the scouting module
-- Overlord and expo ling scouts now get reassigned a new scouting location instead of being cleared. Suicide ling scouts still get cleared (has some benefits early game).

* ScoutingManager v1.11

-- Added expo_positions_complete to the Map_Inventory
-- The scouting expo-ling now avoids the enemy main base and all starting positions, mostly

* Minor Stock Fix

Fixed error where Sum of max stocks were not adding up.

* Player Log Update Pull Request (#128)

* Detection Attempt

The added piece of code is an attempt at detecting the time in seconds of the first enemy detector spotted.

* Detection Addition

This should be able to notify us if we have seen detection units. (Still to be tested on other peoples' bots.

* Detection Update

Refactoring and Code Improvement

* Detection Fix

* Player Data Collection

This update collects the enemy player's data from all the info provided from the Player_Model. Testing has provided results, but they are a little bit sketchy.

* PlayerLogUpdate

Added refence functionality to playerLog function.

* Player Inventory with Description

This is simply the same function but with labelled parameters.

* Player Log fix and neat formatting

Fixed an error in the player log and formatted the output file.

* test

test

* compare

compare

* f

f

* PlayerLog Update

Added new features to the playerLog function, as well as minor other changes.

* Minor Stock Fix

Fixed error where Sum of max stocks were not adding up.

* Fixing BWAP_DIR to enviormental variables.

* Uncommented Read function

Reads data from past game.

* Added comments

Comments were added for clarification of the readPlayerLog() function.

* Finalized Read and Write Functions

Fully operable read and write functions that store and use accurate data pertaining to the player model referenced.

* deleted comment

f

* Extra Curly Bracket

Just a small clean-up.

* Resource Depot Fix (Uncommented)

This is the fixed version of the resource depot counts.

* FAP Update

FAP Update