Superior forecast change
If you lose fodder as a result of a fight, nothing can be done about it even if you retreat, so the lost fodder should not be a factor in defense - it is a sunk cost.  But it should be a factor in choosing to attack, since it determines the damage you do to enemy infrastructure.