Resolved a problem in worker defense code.
Previously, the N workers who were defending were cycled through too quickly, every few frames, resulting in the number of funtional defenders being close to 0. Now, once a worker is a defender, it does not reconsider this status until it is wounded or there is no more reason to fight.