Restored old successful BO
Testing was unsuccessful of previous build.  Main suspicion - Four hatch before pool was a very successful build (the most successful, actually) in my previous bot. The 4HCarapace BO is "better" against truly passive opponents but is so passive and difficult to "snap out" of that it is unresponsive at critical times.   Testing again with the exact original build.