On frame refactor (#240)
This patch is part of a larger refactor.

1) It removes combat simulators from scattered calls to a separate class that acts as a wrapper for FAP.
They are called and then move out of scope. Made a change to air simulations - now consider mutas and hydralisks instead of hack in "flying hydralisks"  for the simulation. To make sure the sim is balanced, I found the least common multiple of the two unit values and hardcoded those amounts.

2) It improves 2 diagnostic functions: a) DiagnosticBar - which draws a segmented bar (like HP) of arbitrary color under a particular unit. b) DiagnosticClock which tells me if a particular function or code segment runs over 15ms.

3) I pull several lines of code out of onFrame() and shift towards McRave's style of each class having an independent thisClass.onFrame() call.

4) It contains 1 play style change - altering the retreating patterns to a shorter, more concise definition. It reduces performance against some bots, but these don't affect the aggregate behavior overall. Once I finish working on surrounds I don't think this should be a factor anymore.