Several Mobility Improvements
Plausibly resolved the mobility challenges within visible motion for now.

Added a "short movement" for air units. I check if I am in a buffer zone or the end destination is under threat. If so, I override the movement and go to safety if "caution" is flagged, otherwise I attack.

Removed most vector math

Installed a fallthrough for getSurroundPosition of least threatened position